Split up canteen test file
